Switched to autotools-based build system, install backend in /usr/lib/print-backends
To not clutter /usr/bin/ with executables which are not intended to be
directly started and also to get an easy overview of which print
dialog backends are installed, we install all print dialog backends in
/usr/lib/print-backends now. This also removes the need to provide a
man page in Debian packaging.

The original build system of a simple Makefile plus a shell script for
"make install" is replaced by a full-featured autotools-based build
system now. This allows for changing destination directories for
installation, makes distribution-packaging easier, and is generally
more flexible.

We also link against the backend library of cpdb-libs instead of
copying the D-Bus interface for backends from there, to make this
package more easily maintainable.